Add 1.5/2.5-bit TurboQuant KV support (outlier channel allocation)
Alternative (slightly shorter): Support fractional 1.5/2.5-bit TurboQuant KV

+++ integrations/vllm_upstream/overlay/vllm/v1/attention/backends/turboquant_attn.py
@@ -177,7 +177,8 @@ class TurboQuantAttentionImpl(AttentionImpl):
             from turboquant import TurboQuantProd
 
             cfg = get_current_vllm_config()
-            bits = int(getattr(cfg.cache_config, "turboquant_bits", 3))
+            # turboquant_bits may be fractional (1.5/2.5) for outlier channel allocation.
+            bits = float(getattr(cfg.cache_config, "turboquant_bits", 3))
             self._quantizer = TurboQuantProd(
                 bits=bits,
                 head_dim=self.head_size,
